Boat placement now reconnects to the game

+++ server/server.js
@@ -496,6 +496,7 @@ const sendPlayerJoin = (ws, playerId) => {
 		JSON.stringify({
 			type: "instruction",
 			instruction: "joinGame",
+			gameId: Game.getGameFromPlayerId(playerId).id,
 			playerId: playerId,
 		}),
 	);
@@ -729,19 +730,18 @@ Deno.serve({ hostname: safeMode ? "localhost" : "0.0.0.0" }, (req) => {
 				case "revealPosition":
 			}
 
-		switch (ev.instruction) {
-			case "createGame":
+		if (ev.type === "lobbyInstruction") {
+			if (ev.instruction === "createGame")
 				return handleCreateGame(ws, ev.playerId, ev?.gameId);
-			case "joinGame":
+			if (ev.instruction === "joinGame")
 				return handleJoinGame(ws, ev.gameId, ev.playerId);
-			case "leaveGame":
+			if (ev.instruction === "leaveGame")
 				return handleLeaveGame(ws, ev.gameId, ev.playerId);
-			case "deleteGame":
+			if (ev.instruction === "deleteGame")
 				return handleDeleteGame(ws, ev.gameId, ev.playerId);
-			case "startGame":
+			if (ev.instruction === "startGame")
 				return startGame(ws, ev.gameId, ev.playerId);
-			default:
-				return sendError(ws, "malformed instruction");
+			return sendError(ws, "malformed instruction");
 		}
 	});
